new file:   src/common/pipe/resize-image.pipe.ts
