+++ src/goods/goods.module.ts
@@ -11,7 +11,7 @@ import { StocksController } from './stocks.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Orders } from '../orders/entities/orders.entity';
 import { Like } from '../like/entities/like.entity';
-import { S3FileService } from '../configs/s3_fileupload';
+import { S3FileService } from '../common/utils/s3_fileupload';
 
 @Module({
   imports: [
